+++ tests/test_optional.py
@@ -25,7 +25,7 @@ def test_optional_dict():
                 c.item("key1") * 500, skip_if=c.item("key1") < 5
             ): c.optional(c.item("key22"), skip_value=20),
         }
-    ).gen_converter(debug=True)
+    ).gen_converter(debug=False)
     assert conv([{"key1": 1, "key2": 2}, {"key1": 10, "key22": 20}], x=1) == [
         {"key1": 1, "key2": 2, "key3": 200, 2: 0},
         {"key1": 10, "key4": 3000, "key5": 3000, 4000: 20},
@@ -47,7 +47,7 @@ def test_optional_list_tuple_set():
             c.optional(c.item("key1") * 2, skip_value=20),
             c.optional(c.item("key1") * 3, skip_if=c.item("key1") < 5),
         ]
-    ).gen_converter(debug=True)
+    ).gen_converter(debug=False)
     assert conv([{"key1": 1, "key2": 2}, {"key1": 10, "key22": 20}]) == [
         [1, 2, 2],
         [10, 30],
@@ -59,7 +59,7 @@ def test_optional_list_tuple_set():
             c.optional(c.item("key1") * 2, skip_value=20),
             c.optional(c.item("key1") * 3, skip_if=c.item("key1") < 5),
         )
-    ).gen_converter(debug=True)
+    ).gen_converter(debug=False)
     assert conv([{"key1": 1, "key2": 2}, {"key1": 10, "key22": 20}]) == [
         (1, 2, 2),
         (10, 30),
@@ -71,7 +71,7 @@ def test_optional_list_tuple_set():
             c.optional(c.item("key1") * 2, skip_value=20),
             c.optional(c.item("key1") * 3, skip_if=c.item("key1") < 5),
         }
-    ).gen_converter(debug=True)
+    ).gen_converter(debug=False)
     assert conv([{"key1": 1, "key2": 2}, {"key1": 10, "key22": 20}]) == [
         {1, 2, 2},
         {10, 30},